+++ packages/components/fields/date-range-field/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @commercetools-uikit/date-range-field
 
+## 19.6.0
+
+### Patch Changes
+
+- Updated dependencies [[`f09213e`](https://github.com/commercetools/ui-kit/commit/f09213e1836b68fef1982503adabb3576f195a54), [`061fe2a`](https://github.com/commercetools/ui-kit/commit/061fe2a32d78a7b845314735ef5de528c8104d0a)]:
+  - @commercetools-uikit/design-system@19.6.0
+  - @commercetools-uikit/field-label@19.6.0
+  - @commercetools-uikit/date-range-input@19.6.0
+  - @commercetools-uikit/constraints@19.6.0
+  - @commercetools-uikit/spacings@19.6.0
+  - @commercetools-uikit/field-errors@19.6.0
+  - @commercetools-uikit/field-warnings@19.6.0
+  - @commercetools-uikit/utils@19.6.0
+
 ## 19.5.0
 
 ### Patch Changes