+++ packages/components/fields/multiline-text-field/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @commercetools-uikit/multiline-text-field
 
+## 19.6.0
+
+### Patch Changes
+
+- Updated dependencies [[`f09213e`](https://github.com/commercetools/ui-kit/commit/f09213e1836b68fef1982503adabb3576f195a54), [`061fe2a`](https://github.com/commercetools/ui-kit/commit/061fe2a32d78a7b845314735ef5de528c8104d0a), [`3df5448`](https://github.com/commercetools/ui-kit/commit/3df5448add57a4b83529ac8503bc836f2e7ced9e)]:
+  - @commercetools-uikit/design-system@19.6.0
+  - @commercetools-uikit/multiline-text-input@19.6.0
+  - @commercetools-uikit/field-label@19.6.0
+  - @commercetools-uikit/constraints@19.6.0
+  - @commercetools-uikit/spacings@19.6.0
+  - @commercetools-uikit/field-errors@19.6.0
+  - @commercetools-uikit/field-warnings@19.6.0
+  - @commercetools-uikit/utils@19.6.0
+
 ## 19.5.0
 
 ### Patch Changes